Messages endpoints (beginning)

+++ src/main/java/rocks/massi/data/GroupsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GroupsRepository extends JpaRepository<Group, Integer> {
+public interface GroupsRepository extends JpaRepository<Group, Long> {
     List<Group> findByName(String name);
 
     List<Group> findByStatus(Group.GroupStatus status);
